Move map to separate component
